Change get_agent_position to not return None (#2343)
If we can't figure out the agents positions, that should be an error. If
there's a reason someone wants to write a test where they're checking
the position of an agent that doesn't exist, they can catch that error.
But for all the upstanding citizens that only ask about agents that do
exist, you now don't need to guard against this returning None.

[Asana
Task](https://app.asana.com/1/1209016784099267/project/1210348820405981/task/1211164822971260)

Co-authored-by: S. Alex Smith <sasmith@stem.ai>